+++ source/ProcessManager.Orchestrations.Tests/Fixtures/OrchestrationsAppFixture.cs
@@ -89,6 +89,10 @@ public class OrchestrationsAppFixture : IAsyncLifetime
             OrchestrationsAppManager.TestLogger,
             IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace,
             IntegrationTestConfiguration.Credential);
+        EnqueueBrs024ServiceBusListener = new ServiceBusListenerMock(
+            OrchestrationsAppManager.TestLogger,
+            IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace,
+            IntegrationTestConfiguration.Credential);
         EnqueueBrs026ServiceBusListener = new ServiceBusListenerMock(
             OrchestrationsAppManager.TestLogger,
             IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace,
@@ -119,6 +123,8 @@ public class OrchestrationsAppFixture : IAsyncLifetime
     public ServiceBusListenerMock EnqueueBrs021ForwardMeteredDataServiceBusListener { get; }
 
     public ServiceBusListenerMock EnqueueBrs023027ServiceBusListener { get; }
+
+    public ServiceBusListenerMock EnqueueBrs024ServiceBusListener { get; }
 
     public ServiceBusListenerMock EnqueueBrs026ServiceBusListener { get; }
 
@@ -181,6 +187,9 @@ public class OrchestrationsAppFixture : IAsyncLifetime
         await EnqueueBrs023027ServiceBusListener.AddTopicSubscriptionListenerAsync(
             ediEnqueueTopicResources.Brs023027Subscription.TopicName,
             ediEnqueueTopicResources.Brs023027Subscription.SubscriptionName);
+        await EnqueueBrs024ServiceBusListener.AddTopicSubscriptionListenerAsync(
+            ediEnqueueTopicResources.Brs024Subscription.TopicName,
+            ediEnqueueTopicResources.Brs024Subscription.SubscriptionName);
         await EnqueueBrs026ServiceBusListener.AddTopicSubscriptionListenerAsync(
             ediEnqueueTopicResources.Brs026Subscription.TopicName,
             ediEnqueueTopicResources.Brs026Subscription.SubscriptionName);
@@ -225,6 +234,7 @@ public class OrchestrationsAppFixture : IAsyncLifetime
         AzuriteManager.Dispose();
         await ServiceBusResourceProvider.DisposeAsync();
         await EnqueueBrs023027ServiceBusListener.DisposeAsync();
+        await EnqueueBrs024ServiceBusListener.DisposeAsync();
         await EnqueueBrs026ServiceBusListener.DisposeAsync();
         await EnqueueBrs028ServiceBusListener.DisposeAsync();
         await IntegrationEventServiceBusListener.DisposeAsync();